+++ py/make_state_plots.py
@@ -27,18 +27,17 @@ def get_nation_weekly(df, pop):
     """
     ndays=7
     l = len(df)
-    nc=np.zeros(l)
-    nd=np.zeros(l)
     deaths = df.deaths
-
-    for i in range(1,l):
-        di = deaths[i]
-        di1 = deaths[i-1]
-        nd[i] = di - di1
-    nd=100000*nd/(ndays*pop)
-    
     dates = df.index.get_level_values('date')
-    return dates, nd
+    nd =[]
+    dates2 =[]
+    for i in range(ndays-1,l,ndays):
+        di = deaths[i]
+        di1 = deaths[i-ndays]
+        nd.append(100000*(di - di1)/(ndays*pop))
+        dates2.append(dates[i])
+        # print(di,di1,di-di1)
+    return dates2, nd
 
 def get_state_weekly(df, pop = None):
     """ Generate lists or arrays of x and y values for a single nation based on daily change averaged over seven days
@@ -52,22 +51,22 @@ def get_state_weekly(df, pop = None):
         Array of average daily deaths per 100,000 people
     """
     if pop is None:
-        pop = df.population
-
+        pop = df.population.iloc[0]
+        if pop == 0:
+            print('hello')
     ndays=7
     l = len(df)
-    nc=np.zeros(l)
-    nd=np.zeros(l)
     deaths = df.deaths
-
-    for i in range(1,l):
-        di = deaths[i]
-        di1 = deaths[i-1]
-        nd[i] = di - di1
-    nd=100000*nd/(ndays*pop)
-    
     dates = df.index.get_level_values('date')
-    return dates, nd
+    nd =[]
+    dates2 =[]
+    for i in range(ndays-1,l,ndays):
+        di = deaths[i]
+        di1 = deaths[i-ndays]
+        nd.append(100000*(di - di1)/(ndays*pop))
+        dates2.append(dates[i])
+        # print(di,di1,di-di1)
+    return dates2, nd
 
 config = get_config()
 
@@ -85,12 +84,7 @@ dmax = df.index.get_level_values('date').max()
 #get us data
 df_us=df[df.index.get_level_values('ISO_A3')=='USA']
 df_us=df_us.groupby(axis='index', by=['date']).sum()
-df_us=df_us.resample('W', level='date', closed='right').last()
-
-#skip partial week at end
-y=df_us[:len(df_us)-1].index.get_level_values('date').max()
-if dmax-y != np.timedelta64(24*7,'h'):
-    df_us=df_us[:len(df_us)-1]
+df_us = df_us[df_us.index.get_level_values('date') > dmax-np.timedelta64(int(config['PLOT CONFIGURATION']['calendar_window'])*7*24,'h')]
 
 #populations of states
 pop = pops_dict['USA']['population']
@@ -107,40 +101,34 @@ for fips in states_fips_s:
         #get weekly data
         df_state=df_states[df_states.index.get_level_values('state_fips') ==fips]
         state = df_state.index.get_level_values('state')[0]
-        df_state=df_state.resample('W', level='date', closed='right').last()
+        df_state = df_state[df_state.index.get_level_values('date') > dmax-np.timedelta64(int(config['PLOT CONFIGURATION']['calendar_window'])*7*24,'h')]
 
-
-        #Avoid partial weeks
-        y=df_state[:len(df_state)-1].index.get_level_values('date').max()
-        if dmax-y != np.timedelta64(24*7,'h'):
-            df_state=df_state[:len(df_state)-1]
-            
         #weekly changes
-        dates_n,nd_state = get_state_weekly(df_state)
+        if state in pops_dict.keys():
+            dates_n,nd_state = get_state_weekly(df_state, int(pops_dict[state]))
 
-        #make plot
+            #make plot
+            fig, ax=plt.subplots()
+            for tick in ax.get_xticklabels():
+                tick.set_rotation(45)
+            ax.plot(dates_n, nd_state)
+            ax.plot(dates, nd)
+            ax.set_ylim(0, MAX_Y)
+            ax.legend([state, 'USA'])
+            ax.set_title('Daily New Fatalities per 100,000 Population')
 
-        fig, ax=plt.subplots()
-        for tick in ax.get_xticklabels():
-            tick.set_rotation(45)
-        ax.plot(dates_n, nd_state)
-        ax.plot(dates, nd)
-        ax.set_ylim(0, MAX_Y)
-        ax.legend([state, 'USA'])
-        ax.set_title('Daily New Fatalities per 100,000 Population')
+            #Put text showing last date and last value
+            last = len(nd_state)-1
+            last_date=f'{dates_n[last]}'[:10]
+            ax.annotate(f'{last_date}, {round(nd_state[last],4)}', [dates_n[last],nd_state[last]])
+            fig.tight_layout(pad=4)
 
-        #Put text showing last date and last value
-        last = len(nd_state)-1
-        last_date=f'{dates_n[last]}'[:10]
-        ax.annotate(f'{last_date}, {round(nd_state[last],4)}', [dates_n[last],nd_state[last]])
-        fig.tight_layout(pad=4)
-
-        #save and upload
-        start = time.time()
-        fig.savefig(config['FILES']['scratch_image'])
-        plt.close()
-        upload_file(config['FILES']['scratch_image'], 'phoenix-technical-services.com', fips + '.jpg', title=fips)
-        os.remove(config['FILES']['scratch_image'])
+            #save and upload
+            start = time.time()
+            fig.savefig(config['FILES']['scratch_image'])
+            plt.close()
+            upload_file(config['FILES']['scratch_image'], 'phoenix-technical-services.com', fips + '.jpg', title=fips)
+            os.remove(config['FILES']['scratch_image'])
 
 ################################################################
 # Canada
@@ -163,14 +151,8 @@ for fips in states_fips_s:
         #get weekly data
         df_state=df_states[df_states.index.get_level_values('state_fips') ==fips]
         state = df_state.index.get_level_values('state')[0]
-        df_state=df_state.resample('W', level='date', closed='right').last()
+        df_state = df_state[df_state.index.get_level_values('date') > dmax-np.timedelta64(int(config['PLOT CONFIGURATION']['calendar_window'])*7*24,'h')]
 
-
-        #Avoid partial weeks
-        y=df_state[:len(df_state)-1].index.get_level_values('date').max()
-        if dmax-y != np.timedelta64(24*7,'h'):
-            df_state=df_state[:len(df_state)-1]
-            
         #weekly changes
         if fips in canada_pop_dict.keys():
             pop = int(canada_pop_dict[fips])
